env-manager/pyproject.toml
[project]
name = "env-manager"
version = "0.1.0"
description = "Environment-aware configuration loader with GCP Secret Manager support"
authors = [{ name = "Bastian Ibañez", email = "bastian.miba@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv>=1.0.0,<2.0.0",
    "pyyaml>=6.0.1,<7.0.0",
    "google-cloud-secret-manager>=2.20.0,<3.0.0",
    "dev-utils @ git+https://github.com/NotoriosTI/libraries.git@main#subdirectory=dev-utils",
    "pytest>=8.0.0,<9.0.0",
    "pytest-mock>=3.14.0,<4.0.0",
]

[tool.poetry]
packages = [{ include = "env_manager", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
